+++ book/snippets/python/src/hello.py
@@ -1,14 +1,15 @@
 # ANCHOR: full
 from nostr_sdk import Keys, Client, EventBuilder
 
-async def hello(relay_url: str):
+
+async def hello():
     # ANCHOR: client
     keys = Keys.generate()
     client = Client(keys)
     # ANCHOR_END: client
 
     # ANCHOR: connect
-    await client.add_relay(relay_url) # For example: "wss://relay.damus.io"
+    await client.add_relay("wss://relay.damus.io")
 
     await client.connect()
     # ANCHOR_END: connect